+++ changelog.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * "View inventory" page unit tests.
     * Add new inventory item feature test.
     * Patch inventory item feature test.
+    * "Edit inventory" page unit tests.
 * Password resetting
     * Added authorization for users.
 * Validation